attack-requirements.txt
black==19.3b0
tlslite-ng==0.7.5
rdtsc==0.2.1
sympy==1.4
